// desktop/src/open-api/model/chartGrouping.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ChartGrouping = 'CATEGORIES' | 'TAGS' | 'PAIDBY';

export const ChartGrouping = {
    Categories: 'CATEGORIES' as ChartGrouping,
    Tags: 'TAGS' as ChartGrouping,
    Paidby: 'PAIDBY' as ChartGrouping
};