// desktop/src/open-api/model/pieChartData.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PieChartDataPoint } from './pieChartDataPoint';


export interface PieChartData { 
    /**
     * Array of pie chart data points
     */
    data: Array<PieChartDataPoint>;
}