// desktop/src/open-api/model/pieChartDataCommand.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ChartGrouping } from './chartGrouping';
import { ReceiptPagedRequestFilter } from './receiptPagedRequestFilter';


export interface PieChartDataCommand { 
    /**
     * What to group the pie chart by
     */
    chartGrouping: ChartGrouping;
    /**
     * Optional filter for receipts
     */
    filter?: ReceiptPagedRequestFilter;
}
export namespace PieChartDataCommand {
}